refactor: consolidate agent directory structure
- Merge src/lib/agent/ (AgentModel, types) into src/lib/agents/
- Move agent role implementations into src/lib/agents/roles/
- Update all imports to go through the agents barrel index
