+++ src/lib/agents/roles/generic.ts
