+++ src/lib/agents/roles/writer.test.ts
