+++ src/lib/agents/roles/writer.ts
