+++ src/lib/tools/EditorTools.test.ts
@@ -4,7 +4,7 @@ import { describe, it, expect, vi, beforeEach } from "vitest";
 import { EditorTools, createDelegateToSkillHandler } from "./EditorTools";
 import { WorkspaceTools } from "./WorkspaceTools";
 import { saveSkills } from "../skills";
-import type { AgentRunnerFactory } from "../agents/factory";
+import type { AgentRunnerFactory } from "../agents";
 
 describe("EditorTools", () => {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
